Change perturb interface output to just mask output + revert to batch op
The first part of splitting out the common operation of mask application
to imagery is to reduce the operation of the PerturbImage interface to
just generating masks.
This included reverting the "streaming" aspect of the interface back to
batch operation as the streaming functionality was overwhelmingly
applicable to the `(image, mask)->image` component of the operation.
Without that responsibility, perturbation implementation are much ligher
and simpler, for which streaming makes less sense currently.
Future use-cases and issues may re-motivate a streaming interface, but
for now they are not known, so simplicity rules.
Additionally, the saliency map generation algorithms take in a batch
matrix of masks, so this new form will match that directly.