Added diagram for tool, minor cleanup, "exercise" at end
Added a simple sequence diagram for the utility and added a few more
words to some explanations.

In order to accommodate CI execution, the parameterization needed to be
lowered to fit below 7GB of RAM usage available to free CI runners. An
"additional example" as previously demoed as been adapted to be an
"exercise for the user".